requirements.txt
aniso8601==9.0.1
azure-core==1.26.3
azure-storage-blob==12.15.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
dnspython==2.3.0
filelock==3.10.2
Flask==2.2.2
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.3
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.0
numpy==1.24.1
opencv-python==4.6.0.66
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.18.0
shapely==2.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.0
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
sympy==1.11.1
tesseract==0.1.3
thinc==8.1.7
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
ultralytics==8.0.61
urllib3==1.26.14
wasabi==1.1.1
Werkzeug==2.2.2
zipp==3.11.0
